Refactoring
Aus der Klasse NeuerAuftrag wurden zwei Models: AuftragModel und
KundeModel. Beide haben zudem einen extra Unterordner im Projekt
bekommen (mehr Models werden wohl folgen)
