+++ Autopilot/Autopilot.csproj
@@ -129,7 +129,8 @@
     <Compile Include="buchung.cs">
       <DependentUpon>EntityDataModel.tt</DependentUpon>
     </Compile>
-    <Compile Include="NeuerAuftrag.cs" />
+    <Compile Include="Models\KundeModel.cs" />
+    <Compile Include="Models\AuftragModel.cs" />
     <Compile Include="EntityDataModel.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
@@ -488,6 +489,7 @@
   <ItemGroup>
     <Service Include="{508349B6-6B84-4DF5-91F0-309BEEBAD82D}" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
